Move WindowHandle to renderer namespace
